Add files via upload
In this report, we begin with a critical examination of salary data across the employees of The Shoppe - a fashion organization based in Nigeria.

We aim to challenge societal norms, ensure inclusivity for all genders, create environments where all genders are valued and respected. By embracing diversity and fostering inclusivity, workplaces can harness the full potential of their employees, driving innovation and success.