Rename VIDEO GAME SALES ANALYSIS - HOPE NKOM.pdf to Video Game Sales Analysis - HOPE NKOM.pdf
The goal of this project is to conduct a comprehensive analysis on all video games sales worldwide. To deep dive into top performing video games, genres, publishers and platform.

The analysis and insights from the Top 20 performing video games, will be instrumental in improving the business model, strategic planning, decision making and overall business performance that will drive revenue and sales growth.

The findings and actionable recommendations is geared towards helping this Video Game sales company navigate the marketing Industry with precision and foresight as well as ensuring competitiveness in the market.